feat(frontend): refactor sign-in components and guard for already sign-in
