+++ src/frontend/src/lib/components/sign-in/AlreadySignedIn.svelte
@@ -1,0 +1,6 @@
+<script lang="ts">
+	import NotFound from '$lib/components/ui/NotFound.svelte';
+	import { i18n } from '$lib/stores/i18n.store';
+</script>
+
+<NotFound warnText={$i18n.sign_in_openid.already_signed_in} />
