+++ src/frontend/src/routes/(home)/+page.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
 	import { fade } from 'svelte/transition';
-	import SignIn from '$lib/components/core/SignIn.svelte';
 	import Launchpad from '$lib/components/launchpad/Launchpad.svelte';
 	import Loaders from '$lib/components/loaders/Loaders.svelte';
+	import SignIn from '$lib/components/sign-in/SignIn.svelte';
 	import { authSignedIn } from '$lib/derived/auth.derived';
 </script>
