+++ src/frontend/src/routes/(standalone)/auth/callback/+page.svelte
@@ -1,62 +1,18 @@
 <script lang="ts">
 	import { onMount } from 'svelte';
-	import { fade } from 'svelte/transition';
-	import { goto } from '$app/navigation';
-	import ContainerCentered from '$lib/components/ui/ContainerCentered.svelte';
-	import Message from '$lib/components/ui/Message.svelte';
-	import Spinner from '$lib/components/ui/Spinner.svelte';
-	import { authenticate } from '$lib/services/auth/auth.openid.services';
-	import { i18n } from '$lib/stores/i18n.store';
+	import AlreadySignedIn from '$lib/components/sign-in/AlreadySignedIn.svelte';
+	import Authenticate from '$lib/components/sign-in/Authenticate.svelte';
+	import { authSignedIn } from '$lib/derived/auth.derived';
 
-	let state = $state<'loading' | 'ok' | 'error'>('loading');
+	let alreadySignIn = $state<boolean | undefined>(undefined);
 
-	const signIn = async () => {
-		const { result } = await authenticate();
-		state = result;
-
-		if (result !== 'ok') {
-			return;
-		}
-
-		setTimeout(async () => await goto('/', { replaceState: true }), 1000);
-	};
-
-	onMount(signIn);
+	onMount(() => {
+		alreadySignIn = $authSignedIn;
+	});
 </script>
 
-<ContainerCentered>
-	{#if state === 'ok'}
-		<Message>
-			{#snippet icon()}
-				<span>👋</span>
-			{/snippet}
-
-			<p>{$i18n.sign_in_openid.success}</p>
-		</Message>
-	{:else if state === 'error'}
-		<Message>
-			{#snippet icon()}
-				<span>❌</span>
-			{/snippet}
-
-			<p>{$i18n.sign_in_openid.error}</p>
-		</Message>
-	{:else}
-		<div in:fade>
-			<Message>
-				{#snippet icon()}
-					<Spinner inline />
-				{/snippet}
-
-				<p>{$i18n.sign_in_openid.in_progress}</p>
-			</Message>
-		</div>
-	{/if}
-</ContainerCentered>
-
-<style lang="scss">
-	span {
-		font-size: var(--font-size-h1);
-		line-height: 1;
-	}
-</style>
+{#if alreadySignIn === true}
+	<AlreadySignedIn />
+{:else if alreadySignIn === false}
+	<Authenticate />
+{/if}
